--- SchoolRush/Assets/Scripts/Kart/CheckpointManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 77d5ad37a7bec4c30a9da0644ec66833
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
