--- SchoolRush/Assets/Scripts/PassengerDemo.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 514a64d337c28fd4c99763600622c802
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
