--- SchoolRush/Assets/Scripts/PassengerDemo/Passenger.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: daeb41513b15dc743a3704a552db5cb7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
