--- SchoolRush/Assets/Scripts/PassengerDemo/PassengerSensor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 07739fd9be2b9fd43addee804c1fed2c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
