--- SchoolRush/Assets/Scripts/PassengerDemo/Traffic.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 96d126a287141754b971178ddf73ee39
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
