--- SchoolRush/Assets/Scripts/PassengerDemo/TrafficSensor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9428201fb7283cc4eaeec03a9b5502de
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
